requirements.txt
spotipy
discord.py
wolframalpha
praw
beautifulsoup4